git-svn-id: http://svn.boost.org/svn/boost/branches/release@63560 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ libs/fusion/test/sequence/adapt_tpl_struct.cpp
@@ -97,14 +97,14 @@ main()
     {
         // conversion from point to vector
         point p = {5, 3};
-        vector<int, short> v(p);
+        vector<int, long> v(p);
         v = p;
     }
 
     {
         // conversion from point to list
         point p = {5, 3};
-        list<int, short> l(p);
+        list<int, long> l(p);
         l = p;
     }
